+++ piff/__init__.py
@@ -83,7 +83,7 @@ from .interp import Interp
 from .mean_interp import Mean
 from .polynomial_interp import Polynomial, polynomial_types
 from .basis_interp import BasisInterp, BasisPolynomial
-from .knn_interp import kNNInterp
+from .knn_interp import KNNInterp
 from .gp_interp import GPInterp
 
 # Outlier handlers are named BlahOutliers where Blah is what they are called in teh config file