+++ piff/gp_interp.py
@@ -19,7 +19,7 @@
 import numpy as np
 import warnings
 import copy
-
+import galsim
 import treegp
 
 from .interp import Interp
@@ -101,6 +101,7 @@ class GPInterp(Interp):
                          [defaulf: 4]
     :param logger:       A logger object for logging debug info. [default: None]
     """
+    _type_name = 'GP'
 
     # treegp currently uses slightly different names for these
     treegp_alias = {
@@ -353,3 +354,14 @@ class GPInterp(Interp):
             gp.initialize(self._X, self._y[:,i], y_err=self._y_err[:,i])
             self.gps.append(gp)
 
+class GaussianProcess(GPInterp):
+    # An alternate name for GPInterp for people who like clearer, more verbose names
+    _type_name = 'GaussianProcess'
+
+class GPInterpDepr(GPInterp):
+    _type_name = 'GPInterp'
+
+    def __init__(self, *args, logger=None, **kwargs):
+        logger = galsim.config.LoggerWrapper(logger)
+        logger.error("WARNING: The name GPInterp is deprecated. Use GP or GaussianProcess instead.")
+        super().__init__(*args, logger=logger, **kwargs)